refactor: 패키지 구조 전부 도메인을 향하도록 변경
