+++ .gitignore
@@ -129,3 +129,5 @@ gradle-app.setting
 *.hprof
 
 # End of https://www.toptal.com/developers/gitignore/api/macos,gradle,intellij+all,visualstudiocode
+
+!*/**/out/
