+++ auth/auth-controller/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
-    implementation project(':auth:auth-service')
+    implementation project(':auth:auth-domain')
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
 }
