+++ auth/auth-domain/src/main/java/shop/woowasap/auth/domain/in/UserCreateUseCase.java
@@ -1,0 +1,9 @@
+package shop.woowasap.auth.domain.in;
+
+import shop.woowasap.auth.domain.in.request.UserCreateRequest;
+
+public interface UserCreateUseCase {
+
+    void createUser(UserCreateRequest userCreateRequest);
+
+}
