--- auth/auth-service/src/main/java/shop/woowasap/auth/service/api/UserCreateUseCase.java
@@ -1,9 +1,0 @@
-package shop.woowasap.auth.service.api;
-
-import shop.woowasap.auth.service.dto.request.UserCreateRequest;
-
-public interface UserCreateUseCase {
-
-    void createUser(UserCreateRequest userCreateRequest);
-
-}
